Ajuste no bug na exclusão de filmes
Ao excluir um filme o código tentava buscar ele depois de ser excluido, o que fazia notificar um erro de filme nao encontrado

+++ src/api/filmes/deleteMovie.php
@@ -2,7 +2,7 @@
 header("Content-Type: application/json");
 include '../db_connection.php';
 
-if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
+if ($_SERVER['REQUEST_METHOD'] === 'DELETE' && isset($_GET['id_filme'])) {
     $id_filme = isset($_GET['id_filme']) ? intval($_GET['id_filme']) : null;
 
     if ($id_filme === null) {
@@ -11,17 +11,29 @@ if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
         exit;
     }
 
-    $delete_sql = "DELETE FROM filmes WHERE id_filme = ?";
-    $stmt = $conn->prepare($delete_sql);
-    $stmt->bind_param("i", $id_filme);
-    $stmt->execute();
+    $check_sql = "SELECT * FROM filmes WHERE id_filme = ?";
+    $check_stmt = $conn->prepare($check_sql);
+    $check_stmt->bind_param("i", $id_filme);
+    $check_stmt->execute();
+    $check_result = $check_stmt->get_result();
 
-    if ($stmt->affected_rows > 0) {
-        echo json_encode(['Filme excluid com sucesso']);
-        http_response_code(204); 
-    } else {
+    if ($check_result->num_rows === 0) {
         echo json_encode(['error' => 'Filme não encontrado']);
         http_response_code(404); 
+        exit;
+    }
+
+    $delete_sql = "DELETE FROM filmes WHERE id_filme = ?";
+    $delete_stmt = $conn->prepare($delete_sql);
+    $delete_stmt->bind_param("i", $id_filme);
+    $delete_stmt->execute();
+
+    if ($delete_stmt->affected_rows > 0) {
+        echo json_encode(['success' => 'Filme excluído com sucesso']);
+        http_response_code(204); 
+    } else {
+        echo json_encode(['error' => 'Erro ao excluir o filme']);
+        http_response_code(500); 
     }
 } elseif (isset($_GET['id_filme'])) {
     $id_filme = isset($_GET['id_filme']) ? intval($_GET['id_filme']) : null;
